qt: Move profile manager to own UI tab
